requirements.txt
bitcoin==1.1.42
Django==1.11
python-dateutil==2.8.0
requests==2.21.0
